Refactored image picking logic and fix import typo in API service file name.
